Share the status and date to a new component
