+++ js/src/components/run/registry.ts
@@ -20,13 +20,12 @@ import { RunFormProps, RunResultViewProps } from "./types";
 import { IconType } from "react-icons";
 import { LuBarChartHorizontalBig } from "react-icons/lu";
 import { MdFormatListNumberedRtl, MdSchema } from "react-icons/md";
+import { As, ComponentWithAs } from "@chakra-ui/react";
 
 interface RegistryEntry<PT, RT, VO = any> {
   title: string;
   icon: IconType;
-  RunResultView?: React.ForwardRefExoticComponent<
-    React.RefAttributes<HTMLDivElement> & RunResultViewProps<PT, RT, VO>
-  >;
+  RunResultView?: ComponentWithAs<As, RunResultViewProps<PT, RT, VO>>;
   RunForm?: React.ComponentType<RunFormProps<PT>>;
 }
